Breaking-up controllers and nav changes
Broke out the tree-related pages to the TreeController, references pages to ReferencesController and added the new CoverageController.
