+++ app/views/coverage.scala.html
@@ -1,0 +1,10 @@
+@import org.webjars.play.WebJarsUtil
+
+@()(implicit webJarsUtil: WebJarsUtil)
+
+@main("Coverage Benchmarks") {
+    <main class="container">
+        <h1>Coverage Benchmarks</h1>
+        <p>Coming soon!</p>
+    </main>
+}
